Refactor for name change
