+++ pyproject.toml
@@ -22,6 +22,7 @@ python = "^3.11"
 requests = "^2.32.3"
 fire = "^0.7.0"
 rich = "^13.9.2"
+pydantic = "^2.9.2"
 
 [build-system]
 requires = ["poetry-core"]
